requirements.txt
APScheduler==3.11.1
fastapi==0.121.0
motor==3.7.1
bcrypt==4.0.1
passlib==1.7.4
pydantic==2.12.4
pydantic-settings==2.11.0
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
redis==7.1.0
spotipy==2.25.2
uvicorn==0.38.0
